+++ MAINTAINERS.yml
@@ -734,7 +734,7 @@ Documentation:
         - jukkar
     files:
         - drivers/ptp_clock/
-        - include/ptp_clock.h
+        - include/drivers/ptp_clock.h
     labels:
         - "area: Clocks"
